BACKLOG-17184: convert image editor to modal
