--- src/javascript/JContent/ImageEditor/index.js
@@ -1,3 +1,0 @@
-import ImageEditor from './ImageEditor.container';
-
-export default ImageEditor;
