+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ConfirmSaveDialog.jsx
