+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/CropPanel/CropPanel.jsx
