+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/CropPanel/index.js
