+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/Feedback.jsx
