+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorActions/ImageEditorActions.test.jsx
