+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorActions/index.js
