+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorDialog.gql-mutations.jsx
