+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorDialog.scss
@@ -25,3 +25,7 @@
 
     background: var(--color-gray_light40);
 }
+
+.content.content {
+    margin-top: var(--spacing-large);
+}
