+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorPreview/ImageEditorPreview.jsx
