+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ImageEditorPreview/index.js
