+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ResizePanel/ResizePanel.jsx
