+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/ResizePanel/index.js
