+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/RotatePanel/RotatePanel.jsx
