+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/RotatePanel/RotatePanel.scss
