+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/RotatePanel/index.js
