+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/SaveAsDialog.scss
