+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/UnsavedChangesDialog/UnsavedChangesDialog.jsx
