+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/UnsavedChangesDialog/index.js
