+++ src/javascript/JContent/actions/editImage/ImageEditorDialog/index.js
@@ -1,0 +1,3 @@
+import ImageEditorDialog from './ImageEditorDialog.container';
+
+export default ImageEditorDialog;
