+++ src/javascript/JContent/actions/editImage/index.js
@@ -1,0 +1,1 @@
+export * from './editImageAction';
